# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RiceText 论坛与小说富文本 API
  version: 0.1.0
  description: 文档、章节、用户互动、附件权益和投票均使用共享契约，并由 SQLite 服务持久化实现。
servers:
  - url: http://localhost:8787
    description: 本地开发 API
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: ricetext_session
      description: OIDC 登录后由 Worker 设置的 HttpOnly session cookie。
...
paths:
...
  /api/forum/novels/{novelId}/chapters/sync:
    post:
      operationId: syncNovelChapters
      tags:
        - 论坛业务
      summary: 对比章节内容哈希
      description: |-
        需要文档 owner、ACL editor 或 moderator。客户端提交本地章节清单（含 SHA-256 哈希），服务端对比已存哈希，返回需要上传的章节与已存在（无需上传）的章节 ID。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: novelId
          in: path
          required: true
          description: 章节所属的文档/小说 ID
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
            description: 章节所属的文档/小说 ID
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - cookieSession: []
      requestBody:
        required: true
        description: 请求字段由共享 Zod 契约校验。
        content:
          application/json:
            schema:
              type: object
              properties:
                chapters:
                  default: []
                  maxItems: 10000
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        minLength: 1
                        maxLength: 128
                        pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                      title:
                        type: string
                        minLength: 1
                        maxLength: 500
                      order:
                        type: integer
                        minimum: 0
                        maximum: 9007199254740991
                      hash:
                        type: string
                        minLength: 1
                        maxLength: 128
                    required:
                      - id
                      - title
                      - order
                      - hash
                    additionalProperties: false
              required:
                - chapters
              additionalProperties: false
      responses:
        "200":
          description: 需要更新与已存在的章节 ID。
          content:
            application/json:
              schema:
                type: object
                properties:
                  toUpdate:
                    type: array
                    items:
                      type: string
                      minLength: 1
                      maxLength: 128
                      pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  existing:
                    type: array
                    items:
                      type: string
                      minLength: 1
                      maxLength: 128
                      pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                required:
                  - toUpdate
                  - existing
                additionalProperties: false
        "401":
          description: 未登录或会话已过期。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
      x-implementation-status: implemented
  /api/forum/novels/{novelId}/chapters/{chapterId}:
    put:
      operationId: saveNovelChapter
      tags:
        - 论坛业务
      summary: 保存单个章节
      description: |-
        需要 author 或 moderator。正文经过与文档相同的白名单校验；该章节 baseRevision 过期返回 409。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: novelId
          in: path
          required: true
          description: 路径参数 novelId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
        - name: chapterId
          in: path
          required: true
          description: 路径参数 chapterId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - cookieSession: []
      requestBody:
        required: true
        description: 请求字段由共享 Zod 契约校验。
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 500
                order:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                content:
                  type: object
                  properties:
                    type:
                      type: string
                      const: doc
                    content:
                      default: []
                      maxItems: 20000
                      type: array
                      items:
                        type: object
                        description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                        additionalProperties: true
                  required:
                    - type
                    - content
                  additionalProperties: false
                hash:
                  type: string
                  minLength: 1
                  maxLength: 128
                baseRevision:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
              required:
                - title
                - order
                - content
                - hash
                - baseRevision
              additionalProperties: false
      responses:
        "201":
          description: 保存后的章节版本摘要。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  title:
                    type: string
                    minLength: 1
                    maxLength: 500
                  order:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  revision:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                required:
                  - id
                  - title
                  - order
                  - revision
                additionalProperties: false
        "401":
          description: 未登录或会话已过期。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "403":
          description: 当前身份无编辑权限。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 当前身份无编辑权限。
        "409":
          description: 章节已被其他修改更新。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 章节已被其他修改更新。
        "422":
          description: 章节内容或字段非法。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 章节内容或字段非法。
      x-implementation-status: implemented
...
  /api/forum/chapters:
    get:
      operationId: listChapters
      tags:
        - 论坛业务
      summary: 读取章节目录
      description: |-
        返回当前文档按 order 排序的持久化章节目录。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - {}
        - cookieSession: []
      responses:
        "200":
          description: 按 order 升序的章节。
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          minLength: 1
                          maxLength: 128
                          pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                        title:
                          type: string
                        order:
                          type: integer
                          minimum: 0
                          maximum: 9007199254740991
                        documentId:
                          type: string
                          minLength: 1
                          maxLength: 128
                          pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                        revision:
                          type: integer
                          minimum: 0
                          maximum: 9007199254740991
                        savedAt:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                        hidden:
                          type: boolean
                      required:
                        - id
                        - title
                        - order
                        - documentId
                        - revision
                        - savedAt
                        - hidden
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
      x-implementation-status: implemented
...